Add ProjectConfiguration
What?
=====

This introduces yaml parsing of input to configure a project (e.g. the
application that Unused is being run on).

This is largely based off of the Haskell version, but allows for
categorization of all of application, test, and config files.

Notably, this is currently missing:

* token exact matches
* aliases

This also currently leaves the responsibility of identifying where to
source Yaml files outside of the configuration itself. That may be
pulled back in, but for now, it's the responsibility of the caller to
provide contents to parse.