Introduce simplistic framework hueristics
What?
=====

This introduces a simple set of framework heuristics to classify a
codebase.

For example, because Rails and Phoenix codebases follow convention over
configuration, one might use these hueristics in their unused
configuration:

```
- name: Rails
  matches_if:
    - token_equals: ApplicationController
    - path_starts_with: app/controllers/
    - path_starts_with: app/models/

- name: Phoenix
  matches_if:
    - token_equals: Application
    - token_equals: Endpoint
    - token_equals: Repo
    - token_equals: Router
    - path_equals: mix.exs
```

All `matches_if` assertions need to be `true` for the configuration to
be used. Note that each member in `matches_if` is a discrete key/value
pair; this allows for the same key to be used multiple times. In the
case of a Rails application, we can assert against the directory
structure, where multiple directories need to exist. In the case of
Phoenix, it's common to have a handful of tokens independent of the
directory structure (this is also due to values not currently supporting
wildcard matches – if there were a way to match on `foo` and `foo_web`,
this would be even more reliable).

This does not include any sort of ranking/scoring algorithm, nor does it
support multiple application support (e.g. if an Ember application were
bundled in a Rails application).