Derive Hash for CtagItem; move to `HashSet<CtagItem>`
What?
=====

This updates `CtagItem` to use a `BTreeMap` in order to derive `Hash`.

By deriving `Hash`, this allows `CtagItem` to be used within `HashSet`.
Based on representations of ctag files, while there can exist multiple
tokens, there should not be duplicates of the entirety of a Ctag item
(e.g. token, path, address, and all metadata).

With the shift to Hash, this also updates all signatures from
`Vec<CtagItem>` to a more representative `HashSet<CtagItem>`.