Log misconfigured .unused.yml configuration to stderr
What?
=====

Unused can be configured via .unused.yml and includes low-likelihood
flagging that can be customized by the person using the tool.

This configuration allows for framework patterns to be skipped, helping
to slim down what unused reports as potentially deletable.

Because this can be user-configured, though, and because some of the
filtering can conflict, it allows for logical fallacies that would
render that low-likelihood configuration useless.

First highlighted in https://github.com/unused-code/unused_rs/issues/1,
the example provided is:

```
- name: LiveView
  token_ends_with: Live
  allowed_tokens:
  - mount
  - render
  - terminate
  - handle_params
  - handle_event
  - handle_call
  - handle_info
```

In this case, token_ends_with limits low-likelihood flagging to any
token that ends with `Live`. The second constraint, captured by
`allowed_tokens`, ensures the token is an exact match of any of the
items in the list. Because of these (the exact match, and substring
match), both constraints cannot work as expected with a binary AND
(unless all of the `allowed_tokens` also ended in View, which doesn't
make sense).

At that, we highlight assertions within the low-likelihood configuration
that conflict because of both exact and partial string matching.
Multiple partial string matches are fine, and exact/partial matching is
fine if they're on different values (tokens vs paths).

This does NOT handle this at the `Vec<Assertion>` level, as other areas
of the configuration, like determining how to classify a codebase, rely
on multiple assertions of the same type, and verify all are met somehow
by tokens in the codebase (e.g. you would have assertions to do token
exact matches for `Application` and `Repo`, among others, for a Phoenix
application).

Resolves #4