Allow ignoring defined paths matching a string
What?
=====

This introduces a flag to ignore results where the string provided is a
substring of the token's path.

For example, if a token was defined in
`app/controllers/api/users_controller.rb`, the flag `--ignore
controllers/api` would strip that result out.

This supports multiple flags on the CLI, as well as comma-delimiting
filters.