Introduce a 'doctor' subcommand
What?
=====

This introduces a 'doctor' subcommand which runs a small set of checks
to provide context to the developer about the status of the codebase and
analysis.

This includes information like whether the tags file generated was
generated via Universal Ctags, how many tokens and files are being
processed, and the types of projects made available from configuration.