Parse tags metadata when parsing tags file
What?
=====

In addition to the tags themselves, a tags file holds information
about the program that generated it.

This updates the ctags parser to capture that information in a
structured way.

Given I don't think we can guarantee the order of metadata, it's not
clear if there's a better way to extract the tags info metadata in a
known and repeatable fashion.