Tighten up generate_with_config
What?
=====

This improvement is mostly minor modifications around how the code was
structured. It removes some intermediate steps, resulting in slightly
more terse code that works well as a pipeline.

The one somewhat fancy thing it does do is apply a pattern documented
in https://abramov.io/rust-dropping-things-in-another-thread to manage
Aho-Corasick structure.