Support use in non-git projects
What?
=====

This replaces the current calculation based on files within a codebase
from using `git ls-files` to `ignored` crate behavior.

While the net result should be the same for git-backed directories, for
non-git-backed directories, this should enable `unused` to work.

This also uses a crude approach to attempting to find tags files for
projects not backed by git.

Closes #31