Add `--harsh` to return an exit status of 1 when tokens are found
What?
=====

It is currently difficult to use unused in CI processes (assuming you
want to fail the build if any tokens are found) because, even when
tokens are found, unused returns a 0 exit status.

This introduces a new flag, `--harsh`, which will return an exit status
of 1 if any tokens are found during a typical run (e.g. not during
`unused doctor` and other commands).

This behavior may be inverted in the future such that standard runs
will, by default, return an exit status of 1 without the flag.

Resolves #22